+++ components/check/box.vue
@@ -4,6 +4,7 @@
     :label="label"
     :is-checked="isChecked"
     :disabled="disabled"
+    :icon-position="iconPosition"
     @click.native="$_onClick"
   >
     <slot>{{label}}</slot>
@@ -33,6 +34,10 @@ export default {
     disabled: {
       type: Boolean,
       default: false,
+    },
+    iconPosition: {
+      type: String,
+      default: 'rt',
     },
   },
 
@@ -76,34 +81,3 @@ export default {
   },
 }
 </script>
-
-<style lang="stylus">
-.md-check-box
-  position relative
-  display inline-block
-  text-align center
-  color checkbox-color
-  font-size checkbox-font-size
-  padding v-gap-sm h-gap-md
-  border 1px solid checkbox-border-color
-  border-radius checkbox-border-radius
-  box-sizing border-box
-  overflow hidden
-  &.is-checked
-    color checkbox-active-color
-    border-color checkbox-active-border-color
-    &.is-disabled
-      color checkbox-active-color
-      border-color checkbox-active-border-color
-      opacity 0.6
-  &.is-disabled
-    color checkbox-disabled-color
-    border-color checkbox-disabled-color
-
-  .md-tag
-    position absolute
-    top 0
-    right 0
-    .quarter-bg
-      background-color checkbox-active-color
-</style>