+++ components/check/demo/cases/demo3.vue
@@ -1,10 +1,10 @@
 <template>
   <div class="md-example-child md-example-child-check md-example-child-check-3">
     <md-check-group v-model="insurants">
-      <md-check-box name="self" disabled>自己</md-check-box>
-      <md-check-box name="couple">配偶</md-check-box>
-      <md-check-box name="parent">父母</md-check-box>
-      <md-check-box name="child">子女</md-check-box>
+      <md-check-box icon-position="lt" name="self" disabled>自己</md-check-box>
+      <md-check-box icon-position="rt" name="couple" disabled>配偶</md-check-box>
+      <md-check-box icon-position="lt" name="parent">父母</md-check-box>
+      <md-check-box icon-position="rt" name="child">子女</md-check-box>
     </md-check-group>
 	</div>
 </template>
@@ -23,7 +23,7 @@ export default {
   },
   data() {
     return {
-      insurants: ['self'],
+      insurants: ['self', 'couple'],
     }
   },
 }